+++ lib/omise/all.rb
@@ -1,7 +1,7 @@
 require "omise/account"
 require "omise/balance"
 require "omise/bank_account"
-require "omise/capabilities"
+require "omise/capability"
 require "omise/card"
 require "omise/chain"
 require "omise/charge"
